Standardisation, pip installation and multi file script (#94)
> proper project declaration
> multi file script
> pip installation
> refactor .gitignore
> Changing variable and function names to comply with python's snake-case standard (except for dtos as it affects the JSON output)
> Adding .editorconfig to ensure IDE consistency
