+++ doc/README.md
@@ -1,6 +1,7 @@
 # Table of Content
 
 - [Default Installation](../README.md#installation)
+- [Development Setup](../README.md#development-setup)
 - [NixOS Flake](https://github.com/TamtamHero/fw-fanctrl/tree/packaging/nix/doc/nix-flake.md)
 - [Commands](./commands.md)
 - [Configuration](./configuration.md)
