+++ src/fw_fanctrl/enum/CommandStatus.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class CommandStatus(str, Enum):
+    SUCCESS = "success"
+    ERROR = "error"
