+++ src/fw_fanctrl/enum/OutputFormat.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class OutputFormat(str, Enum):
+    NATURAL = "NATURAL"
+    JSON = "JSON"
