+++ src/fw_fanctrl/exception/JSONException.py
@@ -1,0 +1,2 @@
+class JSONException(Exception):
+    pass
