+++ src/fw_fanctrl/exception/SocketAlreadyRunningException.py
@@ -1,0 +1,2 @@
+class SocketAlreadyRunningException(Exception):
+    pass
