+++ src/fw_fanctrl/exception/SocketCallException.py
@@ -1,0 +1,2 @@
+class SocketCallException(Exception):
+    pass
