+++ src/fw_fanctrl/exception/UnimplementedException.py
@@ -1,0 +1,2 @@
+class UnimplementedException(Exception):
+    pass
