+++ src/fw_fanctrl/exception/UnknownCommandException.py
@@ -1,0 +1,2 @@
+class UnknownCommandException(Exception):
+    pass
